api/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
onnxruntime
opencv-python-headless
